fix pg function test
