+++ ibis-server/resources/function_list/postgres.csv
@@ -33,4 +33,4 @@ scalar,sign,numeric,,numeric,"Sign of number"
 scalar,to_json,json,,boolean,"Convert to JSON"
 scalar,to_number,numeric,,"text,text","Convert string to number"
 scalar,unistr,varchar,,text,"Postgres: Evaluate escaped Unicode characters in the argument"
-scalar,pg_sleep,,,"numeric","Sleep for specified time in seconds"
+scalar,pg_sleep,,,"bigint","Sleep for specified time in seconds"
