+++ ibis-server/tests/routers/v3/connector/postgres/conftest.py
@@ -5,6 +5,7 @@ import pytest
 import sqlalchemy
 from testcontainers.postgres import PostgresContainer
 
+from app.config import get_config
 from tests.conftest import file_path
 
 pytestmark = pytest.mark.postgres
@@ -17,6 +18,9 @@ def pytest_collection_modifyitems(items):
     for item in items:
         if pathlib.Path(item.fspath).is_relative_to(current_file_dir):
             item.add_marker(pytestmark)
+
+
+function_list_path = file_path("../resources/function_list")
 
 
 @pytest.fixture(scope="module")
@@ -56,3 +60,11 @@ def connection_info(postgres: PostgresContainer) -> dict[str, str]:
 def connection_url(connection_info: dict[str, str]):
     info = connection_info
     return f"postgres://{info['user']}:{info['password']}@{info['host']}:{info['port']}/{info['database']}"
+
+
+@pytest.fixture(autouse=True)
+def set_remote_function_list_path():
+    config = get_config()
+    config.set_remote_function_list_path(function_list_path)
+    yield
+    config.set_remote_function_list_path(None)
